# auth-service/swagger.yaml
openapi: 3.0.0
paths:
  /auth/login:
    post:
      description: login api
      operationId: AppController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginDto"
      responses:
        "201":
          description: ""
      summary: login
      tags: &a1
        - Auth
  /auth/signup:
    post:
      description: signup api
      operationId: AppController_signup
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistrationDto"
      responses:
        "201":
          description: ""
      summary: signup
      tags: *a1
  /auth/admin/update:
    post:
      description: User update api only can be used by admin
      operationId: AppController_update
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateDto"
      responses:
        "201":
          description: ""
      summary: user update api
      tags: *a1
info:
  title: API Docs
  description: NestJS Study API description
  version: 1.0.0
  contact: {}
tags: []
servers: []
components:
  schemas:
    LoginDto:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    UserRegistrationDto:
      type: object
      properties:
        username:
          type: string
          description: username
          example: newbieuser
          minLength: 6
          maxLength: 32
        password:
          type: string
          description: password
          example: securePassword123
          minLength: 8
          maxLength: 32
        re_password:
          type: string
          description: re-typed password
          example: securePassword123
          minLength: 8
          maxLength: 32
      required:
        - username
        - password
        - re_password
    UserUpdateDto:
      type: object
      properties:
        username:
          type: string
        role:
          type: number
          description: "role value: USER = 4, OPERATOR = 3, AUDITOR = 2, ADMIN = 1"
          example: 4
        is_active:
          type: boolean
          description: is user active?
          example: true
      required:
        - username
        - role
        - is_active